refactor(IMS/image): move image data_source to services path
