+++ huaweicloud/services/ims/data_source_huaweicloud_images_image.go
@@ -1,4 +1,4 @@
-package huaweicloud
+package ims
 
 import (
 	"regexp"
@@ -176,7 +176,7 @@ func DataSourceImagesImageV2() *schema.Resource {
 // dataSourceImagesImageV2Read performs the image lookup.
 func dataSourceImagesImageV2Read(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	imageClient, err := config.ImageV2Client(GetRegion(d, config))
+	imageClient, err := config.ImageV2Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud image client: %s", err)
 	}
